+++ src/main/util/updateTkoolmvNamagameKitRuntime.ts
@@ -4,28 +4,22 @@ import * as admZip from "adm-zip";
 import * as sh from "shelljs";
 
 export async function updateTkoolmvNamagameKitRuntime(moduleDirPath: string): Promise<void> {
-	const moduleName = "tkoolmv-namagame-kit";
+	const moduleName = "tkoolmv-namagame-runtime";
 	const versionTxtPath = path.join(moduleDirPath, "version.txt");
 	const latestVersion = await getLatestVersion(moduleName);
-	if (!fs.existsSync(versionTxtPath)) {
-		if (!fs.existsSync(moduleDirPath)) {
-			fs.mkdirSync(moduleDirPath, { recursive: true });
-		}
-		await installModule(moduleName, latestVersion, `tkoolmv-namagame-kit-runtime-${latestVersion}.zip`, moduleDirPath);
-		fs.writeFileSync(versionTxtPath, latestVersion);
-		return;
-	}
-	const currentVersion = fs.readFileSync(versionTxtPath).toString().trim();
-	if (currentVersion !== latestVersion) {
+	if (fs.existsSync(versionTxtPath)) {
+		const currentVersion = fs.readFileSync(versionTxtPath).toString().trim();
 		const current = currentVersion.split(".");
 		const latest = latestVersion.split(".");
-		if (current[0] !== latest[0]) {
+		if (currentVersion === latestVersion || current[0] !== latest[0]) {
 			return;
 		}
-		await installModule(moduleName, latestVersion, `tkoolmv-namagame-kit-runtime-${latestVersion}.zip`, moduleDirPath);
-		fs.writeFileSync(versionTxtPath, latestVersion);
-		return;
 	}
+	if (!fs.existsSync(moduleDirPath)) {
+		fs.mkdirSync(moduleDirPath, { recursive: true });
+	}
+	await installModule(moduleName, latestVersion, `tkoolmv-namagame-runtime-${latestVersion}.zip`, moduleDirPath);
+	fs.writeFileSync(versionTxtPath, latestVersion);
 }
 
 async function getLatestVersion(moduleName: string): Promise<string> {